+++ helix-core/src/test/java/org/apache/helix/controller/strategy/TestTopology.java
@@ -33,7 +33,7 @@ import java.util.Map;
  */
 
 public class TestTopology {
-  private static Logger logger = Logger.getLogger(TestAutoRebalanceStrategy.class);
+  private static Logger logger = Logger.getLogger(TestTopology.class);
 
   @Test
   public void testCreateClusterTopology() {
